+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### PaymentSheet
 * [Added] `opensCardScannerAutomatically` configuration
+* [Changed] The form will now highlight incomplete fields when the user taps a disabled "Confirm" button.
 
 ## 24.20.0 2025-08-11